+++ src/main.rs
@@ -2,7 +2,8 @@ use regex::Regex;
 use std::env;
 use std::fs::read_to_string;
 use submitter::{
-    atcoder, codechef, coderun, codeforces, eolymp, kattis, kep, luogu, toph, ucup, yandex,
+    atcoder, codechef, coderun, codeforces, eolymp, kattis, kep, luogu, repovive, toph, ucup,
+    yandex,
 };
 
 fn site_key_from_url(url: &str) -> Option<String> {
@@ -47,6 +48,7 @@ fn short_name_to_site_key(name: &str) -> Option<String> {
         "eolymp" | "eol" => "eolymp.com",
         "coderun" | "cr" => "coderun.yandex.ru",
         "kep" | "kepuz" => "kep.uz",
+        "repovive" | "rv" => "repovive.com",
         _ => return None,
     };
     Some(key.to_string())
@@ -71,6 +73,7 @@ fn do_login(site_key: &str) {
         "eolymp.com" => eolymp::login(),
         "coderun.yandex.ru" => coderun::login(),
         "kep.uz" => kep::login(),
+        "repovive.com" => repovive::login(),
         _ => eprintln!("Unsupported site: {}", site_key),
     }
 }
@@ -133,6 +136,7 @@ fn main() {
         "eolymp.com" => eolymp::submit(url.clone(), language.clone(), source),
         "coderun.yandex.ru" => coderun::submit(url.clone(), language.clone(), source),
         "kep.uz" => kep::submit(url.clone(), language.clone(), source),
+        "repovive.com" => repovive::submit(url.clone(), language.clone(), source),
         _ => println!("Unsupported domain: {}", site_key),
     }
 }